Add error module
